tools interface and init reorganization
